+++ Content.Shared/Weapons/Ranged/Systems/SharedGunSystem.cs
@@ -37,7 +37,6 @@ using Robust.Shared.Timing;
 using Robust.Shared.Utility;
 using Content.Shared.ADT.DNAGunLocker;
 using Content.Shared.Electrocution;
-using Content.Shared.ADT.Crawling.Components; // ADT-Tweak
 
 namespace Content.Shared.Weapons.Ranged.Systems;
